requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
cachetools==5.5.1
celery==5.4.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
cron-descriptor==1.4.5
Django==4.2.18
django-admin-list-filter-dropdown==1.0.3
django-celery-beat==2.7.0
django-celery-results==2.5.1
django-cors-headers==4.6.0
django-extra-fields==3.0.2
django-mathfilters==1.0.0
django-multiselectfield==0.1.13
django-phonenumber-field==8.0.0
django-rest-swagger==2.2.0
django-timezone-field==7.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
drf-yasg==1.21.8
google-api-core==2.24.1
google-api-python-client==2.160.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.66.0
httplib2==0.22.0
idna==3.10
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.5
kombu==5.4.2
MarkupSafe==3.0.2
openapi-codec==1.3.2
packaging==24.2
phonenumbers==8.13.54
pillow==11.1.0
prompt_toolkit==3.0.50
proto-plus==1.26.0
protobuf==5.29.3
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
PyJWT==2.10.1
pyparsing==3.2.1
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
razorpay==1.4.2
requests==2.31.0
rsa==4.9
setuptools==75.8.0
simplejson==3.19.3
six==1.17.0
sqlparse==0.5.3
standard-imghdr==3.13.0
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
